InstanceTunnel: split entity-bound portion off into "Endpoint" class
